Move project commands to CLI root
