+++ datumaro/cli/commands/__init__.py
@@ -1,6 +1,12 @@
-
 # Copyright (C) 2019-2020 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
 
-from . import add, create, explain, export, remove, merge, convert
+# pylint: disable=redefined-builtin
+
+from . import (
+    create, add, remove,
+    explain,
+    export, merge, convert, transform, filter,
+    diff, ediff,
+)
