+++ datumaro/cli/commands/transform.py
@@ -4,4 +4,4 @@
 
 # pylint: disable=unused-import
 
-from ..contexts.source import build_add_parser as build_parser
+from ..contexts.project import build_transform_parser as build_parser
